create abstract base class
