+++ corehq/blobs/fsdb.py
@@ -4,21 +4,19 @@ from __future__ import absolute_import
 import base64
 import errno
 import os
-import re
 import shutil
 import sys
 from hashlib import md5
 from os.path import commonprefix, exists, isabs, isdir, dirname, join, realpath, sep
-from uuid import uuid4
 
 from corehq.blobs import BlobInfo, DEFAULT_BUCKET
 from corehq.blobs.exceptions import BadName, NotFound
+from corehq.blobs.interface import AbstractBlobDB, SAFENAME
 
 CHUNK_SIZE = 4096
-SAFENAME = re.compile("^[a-z0-9_./-]+$", re.IGNORECASE)
 
 
-class FilesystemBlobDB(object):
+class FilesystemBlobDB(AbstractBlobDB):
     """Filesystem storage for large binary data objects
     """
 
@@ -27,20 +25,6 @@ class FilesystemBlobDB(object):
         self.rootdir = rootdir
 
     def put(self, content, basename="", bucket=DEFAULT_BUCKET):
-        """Put a blob in persistent storage
-
-        :param content: A file-like object in binary read mode.
-        :param basename: Optional name from which the blob name will be
-        derived. This is used to make the unique on-disk filename
-        somewhat recognizable.
-        :param bucket: Optional bucket name used to partition blob data
-        in the persistent storage medium. This may be delimited with
-        file path separators. Nested directories will be created for
-        each logical path element, so it must be a valid relative path.
-        :returns: A `BlobInfo` named tuple. The returned object has a
-        `identifier` member that must be used to get or delete the blob. It
-        should not be confused with the optional `basename` parameter.
-        """
         identifier = self.get_identifier(basename)
         path = self.get_path(identifier, bucket)
         dirpath = dirname(path)
@@ -60,28 +44,12 @@ class FilesystemBlobDB(object):
         return BlobInfo(identifier, length, "md5-" + b64digest)
 
     def get(self, identifier, bucket=DEFAULT_BUCKET):
-        """Get a blob
-
-        :param identifier: The identifier of the object to get.
-        :param bucket: Optional bucket name. This must have the same
-        value that was passed to ``put``.
-        :returns: A file-like object in binary read mode. The returned
-        object should be closed when finished reading.
-        """
         path = self.get_path(identifier, bucket)
         if not exists(path):
             raise NotFound(identifier, bucket)
         return open(path, "rb")
 
     def delete(self, identifier=None, bucket=DEFAULT_BUCKET):
-        """Delete a blob
-
-        :param identifier: The identifier of the object to be deleted. The entire
-        bucket will be deleted if this is not specified.
-        :param bucket: Optional bucket name. This must have the same
-        value that was passed to ``put``.
-        :returns: True if the blob was deleted else false.
-        """
         if identifier is None:
             path = safejoin(self.rootdir, bucket)
             remove = shutil.rmtree
@@ -94,23 +62,7 @@ class FilesystemBlobDB(object):
         return True
 
     def copy_blob(self, content, info, bucket):
-        """Copy blob from other blob database
-
-        :param content: File-like blob content object.
-        :param info: `BlobInfo` object.
-        :param bucket: Bucket name.
-        """
         raise NotImplementedError
-
-    @staticmethod
-    def get_identifier(basename):
-        if not basename:
-            return uuid4().hex
-        if SAFENAME.match(basename) and "/" not in basename:
-            prefix = basename
-        else:
-            prefix = "unsafe"
-        return prefix + "." + uuid4().hex
 
     def get_path(self, identifier=None, bucket=DEFAULT_BUCKET):
         bucket_path = safejoin(self.rootdir, bucket)
