+++ corehq/blobs/s3db.py
@@ -1,6 +1,4 @@
 from __future__ import absolute_import
-import re
-from uuid import uuid4
 from contextlib import contextmanager
 
 from corehq.blobs import BlobInfo, DEFAULT_BUCKET
@@ -12,11 +10,12 @@ from botocore.handlers import calculate_md5
 from botocore.exceptions import ClientError
 from botocore.utils import fix_s3_host
 
+from corehq.blobs.interface import AbstractBlobDB, SAFENAME
+
 DEFAULT_S3_BUCKET = "blobdb"
-SAFENAME = re.compile("^[a-z0-9_./-]+$", re.IGNORECASE)
 
 
-class S3BlobDB(object):
+class S3BlobDB(AbstractBlobDB):
 
     def __init__(self, config):
         kwargs = {}
@@ -35,19 +34,6 @@ class S3BlobDB(object):
         self.db.meta.client.meta.events.unregister('before-sign.s3', fix_s3_host)
 
     def put(self, content, basename="", bucket=DEFAULT_BUCKET):
-        """Put a blob in persistent storage
-
-        :param content: A file-like object in binary read mode.
-        :param basename: Optional name from which the blob name will be
-        derived. This is used to make the unique blob name somewhat
-        recognizable.
-        :param bucket: Optional bucket name used to partition blob data
-        in the persistent storage medium. This may be delimited with
-        slashes (/). It must be a valid relative path.
-        :returns: A `BlobInfo` named tuple. The returned object has a
-        `identifier` member that must be used to get or delete the blob. It
-        should not be confused with the optional `basename` parameter.
-        """
         identifier = self.get_identifier(basename)
         path = self.get_path(identifier, bucket)
         params = {"body": content, "headers": {}}
@@ -61,29 +47,12 @@ class S3BlobDB(object):
         return BlobInfo(identifier, obj.content_length, "md5-" + content_md5)
 
     def get(self, identifier, bucket=DEFAULT_BUCKET):
-        """Get a blob
-
-        :param identifier: The identifier of the object to get.
-        :param bucket: Optional bucket name. This must have the same
-        value that was passed to ``put``.
-        :raises: `NotFound` if the object does not exist.
-        :returns: A file-like object in binary read mode. The returned
-        object should be closed when finished reading.
-        """
         path = self.get_path(identifier, bucket)
         with maybe_not_found(throw=NotFound(identifier, bucket)):
             resp = self._s3_bucket().Object(path).get()
         return ClosingContextProxy(resp["Body"])  # body stream
 
     def delete(self, identifier=None, bucket=DEFAULT_BUCKET):
-        """Delete a blob
-
-        :param identifier: The identifier of the object to be deleted. The entire
-        bucket will be deleted if this is not specified.
-        :param bucket: Optional bucket name. This must have the same
-        value that was passed to ``put``.
-        :returns: True if the blob was deleted else false.
-        """
         path = self.get_path(identifier, bucket)
         s3_bucket = self._s3_bucket()
         with maybe_not_found():
@@ -103,12 +72,6 @@ class S3BlobDB(object):
         return False
 
     def copy_blob(self, content, info, bucket):
-        """Copy blob from other blob database
-
-        :param content: File-like blob content object.
-        :param info: `BlobInfo` object.
-        :param bucket: Bucket name.
-        """
         if info.digest and info.digest.startswith("md5-"):
             content_md5 = info.digest[4:]
         else:
@@ -131,16 +94,6 @@ class S3BlobDB(object):
                 self.db.create_bucket(Bucket=self.s3_bucket_name)
             self._s3_bucket_exists = True
         return self.db.Bucket(self.s3_bucket_name)
-
-    @staticmethod
-    def get_identifier(basename):
-        if not basename:
-            return uuid4().hex
-        if SAFENAME.match(basename) and "/" not in basename:
-            prefix = basename
-        else:
-            prefix = "unsafe"
-        return prefix + "." + uuid4().hex
 
     def get_path(self, identifier=None, bucket=DEFAULT_BUCKET):
         if identifier is None:
